CC-35256: Extract RetryingTransport into its own file
Move the RetryingTransport inner class out of ElasticsearchClient into a
top-level package-private class of its own. Pure refactor, no behavioural
change: the class was already static and took all its collaborators through
its constructor, so it had no coupling to ElasticsearchClient instance state.

Extracting it makes the new file explicit in review and lets the async
retry/backoff logic be unit-tested independently. Also drops five imports
from ElasticsearchClient that only the inner class used (ElasticsearchTransport,
Endpoint, TransportOptions, JsonpMapper, CompletableFuture).
